Hartmann V 7.0: Vector Vector Potential
A was made lowercase (to a) to avoid mixup with lift jargon, and converted back to being an actual vector, bringing (I-J) much closer to zero. Finally, the BC nx condition was commented out, causing I and J to finally match and solution matrices to be creatable. A couple other issues/edits with analysis/solution jargon were resolved. This should be "functioning" in terms of able to run fully to 0.2. I'm not sure if it really ought to end at 0.2, however - the program seemed to run quite quickly on my computer, with each of the ~200 timesteps taking only a couple seconds to run.